refactor: improve CheckFQDN with project-scoped validation
Enhance FQDN validation to support project-scoped checks for stricter validation. This prevents FQDN conflicts within the same project while maintaining backward compatibility with global checks.

- Add optional ProjectID parameter to CheckFQDNInput
- Use CheckFQDNExistsForProject when ProjectID is provided
- Fallback to global CheckFQDNExists for backward compatibility
- Improve code formatting and organization in container_handler.go

This provides more accurate validation in multi-container projects.

LP-490

+++ internal/container/application/check_fqdn.go
@@ -40,7 +40,7 @@ func (uc *CheckFQDNUseCase) Execute(ctx context.Context, input CheckFQDNInput) (
 
 	// If ProjectID is provided, use project-scoped check (stricter validation)
 	if input.ProjectID != nil {
-		exists, err = uc.containerRepo.CheckFQDNExistsForProject(ctx, input.FQDN, *input.ProjectID)
+		exists, err = uc.containerRepo.CheckFQDNExistsForProject(ctx, input.FQDN, uint(*input.ProjectID))
 	} else {
 		// Fallback to global check for backward compatibility
 		exists, err = uc.containerRepo.CheckFQDNExists(ctx, input.FQDN)
